+++ sql/emaj--next_version.sql
@@ -5569,10 +5569,10 @@ $_delete_between_marks_group$
 -- delete rows from all log tables (no need to try to delete if v_firstMarkGlobalSeq and v_lastMarkGlobalSeq are equal)
     v_nbTbl = 0;
     IF v_firstMarkGlobalSeq < v_lastMarkGlobalSeq THEN
--- loop on all tables of the group
+-- loop on all tables that belonged to the group at the end of the period
       FOR r_rel IN
           SELECT quote_ident(rel_log_schema) || '.' || quote_ident(rel_log_table) AS log_table_name FROM emaj.emaj_relation
-            WHERE upper_inf(rel_time_range) AND rel_group = v_groupName AND rel_kind = 'r'
+            WHERE rel_group = v_groupName AND rel_kind = 'r' AND rel_time_range @> v_lastMarkTimeId
             ORDER BY rel_priority, rel_schema, rel_tblseq
       LOOP
 -- delete log rows
@@ -5584,9 +5584,9 @@ $_delete_between_marks_group$
       END LOOP;
     END IF;
 -- process emaj_seq_hole content
--- delete all existing holes (if any) between both marks
+-- delete all existing holes (if any) between both marks for tables that belonged to the group at the end of the period
     DELETE FROM emaj.emaj_seq_hole USING emaj.emaj_relation
-      WHERE upper_inf(rel_time_range) AND rel_group = v_groupName AND rel_kind = 'r'
+      WHERE rel_group = v_groupName AND rel_kind = 'r' AND rel_time_range @> v_lastMarkTimeId
         AND rel_schema = sqhl_schema AND rel_tblseq = sqhl_table
         AND sqhl_begin_time_id >= v_firstMarkTimeId AND sqhl_begin_time_id < v_lastMarkTimeId;
 -- create holes representing the deleted logs
@@ -5600,7 +5600,7 @@ $_delete_between_marks_group$
              (SELECT CASE WHEN sequ_is_called THEN sequ_last_val + sequ_increment ELSE sequ_last_val END FROM emaj.emaj_sequence
                 WHERE sequ_schema = rel_log_schema AND sequ_name = rel_log_sequence AND sequ_time_id = v_firstMarkTimeId)
         FROM emaj.emaj_relation
-        WHERE upper_inf(rel_time_range) AND rel_group = v_groupName AND rel_kind = 'r'
+        WHERE rel_group = v_groupName AND rel_kind = 'r' AND rel_time_range @> v_lastMarkTimeId
           AND 0 <
              (SELECT CASE WHEN sequ_is_called THEN sequ_last_val + sequ_increment ELSE sequ_last_val END FROM emaj.emaj_sequence
                 WHERE sequ_schema = rel_log_schema AND sequ_name = rel_log_sequence AND sequ_time_id = v_lastMarkTimeId)
@@ -5610,12 +5610,12 @@ $_delete_between_marks_group$
 -- now the sequences related to the mark to delete can be suppressed
 --   delete first application sequences related data for the group
     DELETE FROM emaj.emaj_sequence USING emaj.emaj_relation
-      WHERE upper_inf(rel_time_range) AND rel_group = v_groupName AND rel_kind = 'S'
+      WHERE rel_group = v_groupName AND rel_kind = 'S' AND rel_time_range @> v_lastMarkTimeId
         AND sequ_schema = rel_schema AND sequ_name = rel_tblseq
         AND sequ_time_id > v_firstMarkTimeId AND sequ_time_id < v_lastMarkTimeId;
 --   delete then emaj sequences related data for the group
     DELETE FROM emaj.emaj_sequence USING emaj.emaj_relation
-      WHERE upper_inf(rel_time_range) AND rel_group = v_groupName AND rel_kind = 'r'
+      WHERE rel_group = v_groupName AND rel_kind = 'r' AND rel_time_range @> v_lastMarkTimeId
         AND sequ_schema = rel_log_schema AND sequ_name = rel_log_sequence
         AND sequ_time_id > v_firstMarkTimeId AND sequ_time_id < v_lastMarkTimeId;
 -- in emaj_mark, reset the mark_logged_rlbk_target_mark column to null for marks of the group that will remain
@@ -5649,7 +5649,15 @@ $emaj_get_consolidable_rollbacks$
       SELECT m1.mark_group AS cons_group,
              m2.mark_name AS cons_target_rlbk_mark_name, m2.mark_id AS cons_target_rlbk_mark_id,
              m1.mark_name AS cons_end_rlbk_mark_name, m1.mark_id AS cons_end_rlbk_mark_id,
-             cast(coalesce(emaj._sum_log_stat_group(m1.mark_group, m2.mark_time_id, m1.mark_time_id),0) AS BIGINT) AS cons_rows,
+             cast(coalesce(
+                  (SELECT sum(emaj._log_stat_tbl(emaj_relation,
+                                                 -- the start mark = max(begin of the rollback time frame, time when the relation has been added to the group)
+                                                 CASE WHEN m2.mark_time_id > lower(rel_time_range) THEN m2.mark_time_id ELSE lower(rel_time_range) END,
+                                                 m1.mark_time_id))
+                     FROM emaj.emaj_relation
+                           -- for tables belonging to the group at the rollback time
+                     WHERE rel_group = m1.mark_group AND rel_kind = 'r' AND rel_time_range @> m1.mark_time_id)
+                          ,0) AS BIGINT) AS cons_rows,
              cast((SELECT count(*) FROM emaj.emaj_mark m3
                    WHERE m3.mark_group = m1.mark_group AND m3.mark_id > m2.mark_id AND m3.mark_id < m1.mark_id) AS INT) AS cons_marks
         FROM emaj.emaj_mark m1